+++ sdk/resourcemanager/eventhub/armeventhub/zz_generated_pagers.go
@@ -16,6 +16,60 @@ import (
 	"reflect"
 )
 
+// ApplicationGroupClientListByNamespacePager provides operations for iterating over paged responses.
+type ApplicationGroupClientListByNamespacePager struct {
+	client    *ApplicationGroupClient
+	current   ApplicationGroupClientListByNamespaceResponse
+	err       error
+	requester func(context.Context) (*policy.Request, error)
+	advancer  func(context.Context, ApplicationGroupClientListByNamespaceResponse) (*policy.Request, error)
+}
+
+// Err returns the last error encountered while paging.
+func (p *ApplicationGroupClientListByNamespacePager) Err() error {
+	return p.err
+}
+
+// NextPage returns true if the pager advanced to the next page.
+// Returns false if there are no more pages or an error occurred.
+func (p *ApplicationGroupClientListByNamespacePager) NextPage(ctx context.Context) bool {
+	var req *policy.Request
+	var err error
+	if !reflect.ValueOf(p.current).IsZero() {
+		if p.current.ApplicationGroupListResult.NextLink == nil || len(*p.current.ApplicationGroupListResult.NextLink) == 0 {
+			return false
+		}
+		req, err = p.advancer(ctx, p.current)
+	} else {
+		req, err = p.requester(ctx)
+	}
+	if err != nil {
+		p.err = err
+		return false
+	}
+	resp, err := p.client.pl.Do(req)
+	if err != nil {
+		p.err = err
+		return false
+	}
+	if !runtime.HasStatusCode(resp, http.StatusOK) {
+		p.err = runtime.NewResponseError(resp)
+		return false
+	}
+	result, err := p.client.listByNamespaceHandleResponse(resp)
+	if err != nil {
+		p.err = err
+		return false
+	}
+	p.current = result
+	return true
+}
+
+// PageResponse returns the current ApplicationGroupClientListByNamespaceResponse page.
+func (p *ApplicationGroupClientListByNamespacePager) PageResponse() ApplicationGroupClientListByNamespaceResponse {
+	return p.current
+}
+
 // ClustersClientListByResourceGroupPager provides operations for iterating over paged responses.
 type ClustersClientListByResourceGroupPager struct {
 	client    *ClustersClient